ml-engine/requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
torch>=2.2.0
redis>=5.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0
joblib>=1.4.0
ta>=0.11.0
